refactor: split ai_player, health, stats into module directories
